🚚 Move biometric to util struc
