+++ packages/starknet_flutter/ios/Classes/Utils/AuthenticationUtil.swift
@@ -1,0 +1,1 @@
+../../../darwin/Classes/Utils/AuthenticationUtil.swift
